reorganizing to consolidate the app_common and app_trapi apps
