+++ PLATER/services/app_trapi.py
@@ -1,22 +1,30 @@
 """FastAPI app."""
 from fastapi import Body, Depends, FastAPI, Response, Request
 from fastapi.encoders import jsonable_encoder
-from fastapi.responses import ORJSONResponse
-from typing import Dict
+from fastapi.responses import ORJSONResponse, RedirectResponse
+from typing import Any, Dict, List
 from pydantic import ValidationError
 
-import orjson
-
-from reasoner_transpiler.exceptions import InvalidPredicateError, InvalidQualifierError, InvalidQualifierValueError, UnsupportedError
-from PLATER.models.shared import ReasonerRequest, MetaKnowledgeGraph, SRITestData
+from reasoner_transpiler.exceptions import (
+    InvalidPredicateError, InvalidQualifierError, InvalidQualifierValueError, UnsupportedError
+)
+from PLATER.models.models_trapi_1_0 import (
+    Message, ReasonerRequest, CypherRequest, SimpleSpecResponse, SimpleSpecElement, CypherResponse
+)
+from PLATER.models.shared import MetaKnowledgeGraph, SRITestData
+from PLATER.services.util.api_utils import (
+    get_graph_interface, get_example, CustomORJSONResponse
+)
+from PLATER.services.util.bl_helper import BLHelper, get_bl_helper
 from PLATER.services.util.graph_adapter import GraphInterface
 from PLATER.services.util.metadata import GraphMetadata
-from PLATER.services.util.question import Question
 from PLATER.services.util.overlay import Overlay
-from PLATER.services.util.api_utils import get_graph_interface, construct_open_api_schema
+from PLATER.services.util.question import Question
 from PLATER.services.config import config
 from PLATER.services.util.logutil import LoggingUtil
-from PLATER.services.app_common import APP_COMMON
+
+
+APP = FastAPI(openapi_url='/openapi.json', docs_url='/docs')
 
 logger = LoggingUtil.init_logging(
     __name__,
@@ -24,10 +32,9 @@ logger = LoggingUtil.init_logging(
     config.get('logging_format'),
 )
 
-# Mount open api at /openapi.json
-APP_TRAPI = APP_COMMON
 
-
+# read in and validate the meta kg json only once on startup,
+# create an already-encoded object that is ready to be returned quickly
 def get_meta_kg_response(graph_metadata_reader: GraphMetadata):
     meta_kg_json = graph_metadata_reader.get_meta_kg()
     try:
@@ -39,9 +46,13 @@ def get_meta_kg_response(graph_metadata_reader: GraphMetadata):
         return None
 
 
+# process and store static objects ready to be returned by their respective endpoints
 graph_metadata_reader = GraphMetadata()
+GRAPH_METADATA = graph_metadata_reader.get_metadata()
 META_KG_RESPONSE = get_meta_kg_response(graph_metadata_reader)
 SRI_TEST_DATA = graph_metadata_reader.get_sri_testing_data()
+
+# get an example query for the /query endpoint, to be included in the open api spec
 TRAPI_QUERY_EXAMPLE = graph_metadata_reader.get_example_qgraph()
 
 
@@ -54,43 +65,53 @@ async def get_meta_knowledge_graph() -> ORJSONResponse:
                               content=META_KG_RESPONSE,
                               media_type="application/json")
     else:
+        # if META_KG_RESPONSE is None it means the meta kg did not validate
         return ORJSONResponse(status_code=500,
                               media_type="application/json",
                               content={"description": "MetaKnowledgeGraph failed validation - "
                                                       "please notify maintainers."})
+
+APP.add_api_route(
+    "/meta_knowledge_graph",
+    get_meta_knowledge_graph,
+    methods=["GET"],
+    response_model=None,
+    responses={200: {"model": MetaKnowledgeGraph}},
+    summary="Meta knowledge graph representation of this TRAPI web service.",
+    description="Returns meta knowledge graph representation of this TRAPI web service.",
+    tags=["trapi"]
+)
 
 
 async def get_sri_testing_data():
     """Handle /sri_testing_data."""
     return SRI_TEST_DATA
 
-
-def orjson_default(obj):
-    if isinstance(obj, set):
-        return list(obj)
-    raise TypeError
-
-
-class CustomORJSONResponse(Response):
-    def render(self, content: dict) -> bytes:
-        return orjson.dumps(content,
-                            default=orjson_default)
+APP.add_api_route(
+    "/sri_testing_data",
+    get_sri_testing_data,
+    methods=["GET"],
+    response_model=SRITestData,
+    response_model_exclude_none=True,
+    summary="Test data for usage by the SRI Testing Harness.",
+    description="Returns a list of edges that are representative examples of the knowledge graph.",
+    tags=["trapi"]
+)
 
 
 async def reasoner_api(
-        response: Response,
         request: ReasonerRequest = Body(
             ...,
             example=TRAPI_QUERY_EXAMPLE,
         ),
-        profile: bool = False,
+        profile: bool = False,  # looks like it's not used, but it's here so that it's documented in the open api spec
         validate: bool = False,
         graph_interface: GraphInterface = Depends(get_graph_interface),
 ) -> CustomORJSONResponse:
 
-    """Handle TRAPI request."""
+    """Handle /query TRAPI request."""
     request_json = request.dict(by_alias=True)
-    # default workflow
+    # use lookup as the default workflow
     workflow = request_json.get('workflow') or [{"id": "lookup"}]
     workflows = {wkfl['id']: wkfl for wkfl in workflow}
 
@@ -119,6 +140,8 @@ async def reasoner_api(
 
     if validate:
         try:
+            # Attempt to parse the request_json using the pydantic model, if it fails it will throw a ValidationError.
+            # Don't save the pydantic object created, we're returning a CustomORJSONResponse instead.
             ReasonerRequest.parse_obj(request_json)
         except ValidationError as e:
             json_response = CustomORJSONResponse(content={"description": f"Validation Errors Occurred: {e.errors()}"},
@@ -130,30 +153,7 @@ async def reasoner_api(
     json_response = CustomORJSONResponse(content=request_json, media_type="application/json")
     return json_response
 
-
-APP_TRAPI.add_api_route(
-    "/meta_knowledge_graph",
-    get_meta_knowledge_graph,
-    methods=["GET"],
-    response_model=None,
-    responses={200: {"model": MetaKnowledgeGraph}},
-    summary="Meta knowledge graph representation of this TRAPI web service.",
-    description="Returns meta knowledge graph representation of this TRAPI web service.",
-    tags=["trapi"]
-)
-
-APP_TRAPI.add_api_route(
-    "/sri_testing_data",
-    get_sri_testing_data,
-    methods=["GET"],
-    response_model=SRITestData,
-    response_model_exclude_none=True,
-    summary="Test data for usage by the SRI Testing Harness.",
-    description="Returns a list of edges that are representative examples of the knowledge graph.",
-    tags=["trapi"]
-)
-
-APP_TRAPI.add_api_route(
+APP.add_api_route(
     "/query",
     reasoner_api,
     methods=["POST"],
@@ -164,13 +164,212 @@ APP_TRAPI.add_api_route(
     tags=["trapi"]
 )
 
-# env var PROFILE_EVERYTHING=true could be used to turn on profiling / speedscope results for all http endpoints
+
+###########################################
+# The following endpoints all come from the old app_common.py file, which was previously a different sub-application.
+###########################################
+
+async def cypher(
+        request: CypherRequest = Body(
+            ...,
+            example={"query": "MATCH (n) RETURN count(n)"},
+        ),
+        graph_interface: GraphInterface = Depends(get_graph_interface),
+) -> CypherResponse:
+    """Handle cypher."""
+    request = request.dict()
+    results = await graph_interface.run_cypher(
+        request["query"],
+        return_errors=True,
+    )
+    return results
+
+
+APP.add_api_route(
+    "/cypher",
+    cypher,
+    methods=["POST"],
+    response_model=CypherResponse,
+    summary="Run cypher query",
+    description=(
+        "Runs cypher query against the Neo4j instance, and returns an "
+        "equivalent response expected from a Neo4j HTTP endpoint "
+        "(https://neo4j.com/docs/rest-docs/current/)."
+    ),
+)
+
+
+async def overlay(
+        request: ReasonerRequest = Body(
+            ...,
+            example={"message": get_example("overlay")},
+        ),
+        graph_interface: GraphInterface = Depends(get_graph_interface),
+) -> Message:
+    """Handle TRAPI request."""
+    overlay_class = Overlay(graph_interface)
+    return await overlay_class.overlay_support_edges(request.dict()["message"])
+
+
+APP.add_api_route(
+    "/overlay",
+    overlay,
+    methods=["POST"],
+    response_model=Message,
+    description=(
+        "Given a ReasonerAPI graph, add support edges for any nodes linked in "
+        "result bindings."
+    ),
+    summary="Overlay results with available connections between each node.",
+    tags=["translator"]
+)
+
+
+async def metadata() -> Any:
+    """Handle /metadata."""
+    return GRAPH_METADATA
+
+APP.add_api_route(
+    "/metadata",
+    metadata,
+    methods=["GET"],
+    response_model=Any,
+    summary="Metadata about the knowledge graph.",
+    description="Returns JSON with metadata about the data sources in this knowledge graph.",
+)
+
+
+async def one_hop(
+        source_type: str,
+        target_type: str,
+        curie: str,
+        graph_interface: GraphInterface = Depends(get_graph_interface),
+) -> List[Dict]:
+    """Handle one-hop."""
+    return await graph_interface.get_single_hops(
+        source_type,
+        target_type,
+        curie,
+    )
+
+APP.add_api_route(
+    "/{source_type}/{target_type}/{curie}",
+    one_hop,
+    methods=["GET"],
+    response_model=List,
+    summary=(
+        "Get one hop results from source type to target type. "
+        "Note: Please GET /predicates to determine what target goes "
+        "with a source"
+    ),
+    description=(
+        "Returns one hop paths from `source_node_type`  with `curie` "
+        "to `target_node_type`."
+    ),
+)
+
+
+async def node(
+        node_type: str,
+        curie: str,
+        graph_interface: GraphInterface = Depends(get_graph_interface),
+) -> List[List[Dict]]:
+    """Handle node lookup."""
+    return await graph_interface.get_node(
+        node_type,
+        curie,
+    )
+
+APP.add_api_route(
+    "/{node_type}/{curie}",
+    node,
+    methods=["GET"],
+    response_model=List,
+    summary="Find `node` by `curie`",
+    description="Returns `node` matching `curie`.",
+)
+
+
+async def simple_spec(
+        source: str = None,
+        target: str = None,
+        graph_interface: GraphInterface = Depends(get_graph_interface),
+        bl_helper: BLHelper = Depends(get_bl_helper),
+) -> SimpleSpecResponse:
+    """Handle simple spec."""
+    source_id = source
+    target_id = target
+    if source_id or target_id:
+        minischema = []
+        mini_schema_raw = await graph_interface.get_mini_schema(
+            source_id,
+            target_id,
+        )
+        for row in mini_schema_raw:
+            source_labels = await bl_helper.get_most_specific_concept(
+                row['source_label']
+            )
+            target_labels = await bl_helper.get_most_specific_concept(
+                row['target_label']
+            )
+            for source_type in source_labels:
+                for target_type in target_labels:
+                    minischema.append((
+                        source_type,
+                        row['predicate'],
+                        target_type,
+                    ))
+        minischema = list(set(minischema))  # remove dups
+        return list(map(lambda x: SimpleSpecElement(**{
+                'source_type': x[0],
+                'target_type': x[2],
+                'edge_type': x[1],
+            }), minischema))
+    else:
+        schema = graph_interface.get_schema()
+        reformatted_schema = []
+        for source_type in schema:
+            for target_type in schema[source_type]:
+                for edge in schema[source_type][target_type]:
+                    reformatted_schema.append(SimpleSpecElement(**{
+                        'source_type': source_type,
+                        'target_type': target_type,
+                        'edge_type': edge
+                    }))
+        return reformatted_schema
+
+APP.add_api_route(
+    "/simple_spec",
+    simple_spec,
+    methods=["GET"],
+    response_model=SimpleSpecResponse,
+    summary="Get one-hop connection schema",
+    description=(
+        "Returns a list of available predicates when choosing a single source "
+        "or target curie. Calling this endpoint with no query parameters will "
+        "return all possible hops for all types."
+    ),
+)
+
+
+async def redirect_to_docs():
+    return RedirectResponse(url="/docs")
+
+APP.add_api_route(
+    "/",
+    redirect_to_docs,
+    include_in_schema=False,
+    methods=["GET"]
+)
+
+# config var PROFILER_ON=true can be used to turn on profiling for all http endpoints
+# even with PROFILER_ON the query param "profile" is what makes an endpoint use profiling and return the results
 if config.get('PROFILER_ON', False) and (config.get('PROFILER_ON') not in ("false", "False")):
     from pyinstrument import Profiler
     from pyinstrument.renderers import SpeedscopeRenderer
     from fastapi.responses import HTMLResponse
 
-    @APP_TRAPI.middleware("http")
+    @APP.middleware("http")
     async def profile_request(request: Request, call_next):
         profiling = request.query_params.get("profile", "false")
         if profiling and profiling != "false":
